refactor: move DateNavigationHeader to common for cross-page reuse
