+++ src/components/common/DateNavigationHeader.tsx
